refactor: move core agent's integration tests
This way it's cleaner and easier to find out where the tests are.
Also the integration tests will compile faster as it avoids double compilation
of the binary targets which are built by the deployer.
