+++ control-plane/agents/core/src/controller/reconciler/volume/garbage_collector.rs
@@ -278,74 +278,3 @@ async fn is_replica_healthy(
         Ok(info.is_replica_healthy(&replica_spec.uuid))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use common_lib::types::v0::openapi::models;
-    use deployer_cluster::ClusterBuilder;
-    use std::time::Duration;
-
-    #[tokio::test]
-    async fn disown_unused_replicas() {
-        const POOL_SIZE_BYTES: u64 = 128 * 1024 * 1024;
-        let reconcile_period = Duration::from_millis(200);
-        let cluster = ClusterBuilder::builder()
-            .with_rest(true)
-            .with_agents(vec!["core"])
-            .with_io_engines(1)
-            .with_tmpfs_pool(POOL_SIZE_BYTES)
-            .with_cache_period("1s")
-            .with_reconcile_period(reconcile_period, reconcile_period)
-            .build()
-            .await
-            .unwrap();
-
-        let rest_api = cluster.rest_v00();
-        let volumes_api = rest_api.volumes_api();
-        let node = cluster.node(0).to_string();
-
-        let volume = volumes_api
-            .put_volume(
-                &"1e3cf927-80c2-47a8-adf0-95c481bdd7b7".parse().unwrap(),
-                models::CreateVolumeBody::new(
-                    models::VolumePolicy::default(),
-                    1,
-                    5242880u64,
-                    false,
-                ),
-            )
-            .await
-            .unwrap();
-
-        let volume = volumes_api
-            .put_volume_target(&volume.spec.uuid, &node, models::VolumeShareProtocol::Nvmf)
-            .await
-            .unwrap();
-
-        cluster.composer().pause(&node).await.unwrap();
-        volumes_api
-            .del_volume_target(&volume.spec.uuid, Some(false))
-            .await
-            .expect_err("io-engine is down");
-        cluster.composer().kill(&node).await.unwrap();
-
-        let volume = volumes_api.get_volume(&volume.spec.uuid).await.unwrap();
-        tracing::info!("Volume: {:?}", volume);
-
-        assert!(volume.spec.target.is_some(), "Unpublish failed");
-
-        let specs = cluster.rest_v00().specs_api().get_specs().await.unwrap();
-        let replica = specs.replicas.first().cloned().unwrap();
-        assert!(replica.owners.volume.is_some());
-        assert!(replica.owners.nexuses.is_empty());
-
-        // allow the reconcile to run - it should not disown the replica
-        tokio::time::sleep(reconcile_period * 12).await;
-
-        let specs = cluster.rest_v00().specs_api().get_specs().await.unwrap();
-        let replica = specs.replicas.first().cloned().unwrap();
-        // we should still be part of the volume
-        assert!(replica.owners.volume.is_some());
-        assert!(replica.owners.nexuses.is_empty());
-    }
-}
