+++ control-plane/agents/core/tests/test.rs
@@ -1,0 +1,6 @@
+mod controller;
+mod nexus;
+mod node;
+mod pool;
+mod volume;
+mod watch;
