+++ k8s/supportability/Cargo.toml
@@ -38,7 +38,6 @@ tar = "0.4"
 flate2 = { version = "1.0.24" }
 chrono = "0.4.19"
 urlencoding = "2.1.0"
-reqwest = { version = "0.11", features = ["json"] }
 downcast-rs = "1.2.0"
 schemars = "0.8.10"
 k8s-operators = { path = "../operators" }
